move special eval logic out of mathscope
