+++ client/src/util/MathScope/index.ts
@@ -4,13 +4,27 @@ import {
   DuplicateAssignmentError,
   UnmetDependencyError,
 } from "./Evaluator";
-import type { MathNode, Parse } from "./interfaces";
+import type {
+  MathNode,
+  EvaluationScope,
+  AnonMathNode,
+  Parse,
+} from "./interfaces";
 import MathScope, { IdentifiedParseable, OnChangeListener } from "./MathScope";
+import { EvaluationError } from "./errors";
 
-export type { IdentifiedParseable, MathNode, OnChangeListener, Parse };
+export type {
+  IdentifiedParseable,
+  MathNode,
+  OnChangeListener,
+  Parse,
+  EvaluationScope,
+  AnonMathNode,
+};
 
 export {
   adapter,
+  EvaluationError,
   AssignmentError,
   DuplicateAssignmentError,
   UnmetDependencyError,
